feat/role: configure jenkins with CASC plugin
Use the 'Configuration as Code' plugin to configure Jenkins.

This commit has a basic password for the initial admin user. It will be
replaced shortly.

There is also a switch here to using the docker_image module for
managing the containers. This requires the Python docker library to be
installed.