feat/role: provide user password via secret
This uses the secrets mechanism to provide the 'chriso' user for Jenkins
with a password. The passwords are encrypted using ansible-vault. The
disadvantage of this mechanism at the moment is that the secrets have to
reside in files on the disk of the host machine, because they have to be
read by docker-compose, which then mounts them into the container. We
can refactor this when Vault is introduced.

There are also a couple of housekeeping corrections in here to have
the compose file directory controlled with a variable.