role/feat: use external java role
Any slaves to connect to Jenkins need to have a Java installation. Each
platform has a slightly different way of installing it, and this
external role seems to cover a lot of cases. I tried to actually set
this up with ansible-galaxy and a requirements file, but I was getting a
strange error when galaxy was running.