refactor/vm: parameterize hostname and IP
This introduces group_vars for defining information related to the hosts
in the Vagrant environment. In the jenkins.yml configuration, the IP
address for the slave is used rather than the host name, just because
Jenkins is running in a container and it would be much more complex to
use the hostname, since that would need to be available inside the
container.