feat/plugin: job DSL plugin
This plugin gives us the ability to define our jobs in code. Note that
pipelines should not be defined using this plugin. Just the actual *job*
that contains the pipeline should be configured using this plugin.

There's also a little utility script here that lists the plugins for the
instance. This is so that you can get the ID and version of the plugin
to add to the plugins.txt file.