fix/build: commit failure on 2nd run
The build was failing on its 2nd run because the command to get the
container ID was assuming that only 1 container would be returned.
However, the grep for 'safe_client_libs_mock' meant it was returning 2
containers, one named 'safe_client_libs_mock' and the other
'safe_client_libs_mock_test_run'.

The solution is to provide the `-w` argument to grep to match an exact
word. There's also another preventative measure to remove the test run
container when it completes, since it's not actually needed beyond that.