refactor/build: more parallel support
Every container that is spun up is appended with the current build
number. It's possible for more than 1 build to be running on any agent
at a given time, so during a clean up you don't want to attempt to
remove containers that are being used by another build.

The clean up also only tries to remove the images committed during that
build. There shouldn't be any stopped container instances that haven't
been removed.