feat/role: configure jenkins slave for Windows
At the moment this is an extremely simple role that installs Java,
creates a working directory for the agent and copies over the agent jar
file for communication with Jenkins.

It needs to be extended to setup the agent to run as a service. The
agent also needs to be started, but I think this might be done in a
separate role, since it can only successfully be launched if Jenkins is
already started and running.