feat/vm: windows server 2016 jenkins slave
This slave will be used for building Safe Client Libs. It's intended to
replicate the Windows environment that Travis have setup:
https://docs.travis-ci.com/user/reference/windows/

The batch file here installs all the same packages from Chocolatey. It
should be noted that it will only be able to install the 'wsl' package
if the Windows updates have been applied to the machine. Earlier builds
of server 2016 don't have the Windows Subsystem for Linux feature
available for installation.