feat/role: run Jenkins slave as Windows service
The Jenkins slave is run as a Windows service. This is setup using
'WinSW', which is a tool I think the Jenkins developers actually
created. It's a wrapper for easily setting up a service. See here for
more details:
https://github.com/kohsuke/winsw

In order to add the slave successfully, you need to obtain a secret key
from the Jenkins master. This is done using a simple Powershell script.
After that's obtained, it can be placed in the configuration file for
the service, and the service is then launched.