feat/vm: box that replicates Travis on Windows
Provides a base box that is intended to mirror the Travis Windows
environment. They have detailed what's in the environment with some
documentation:
https://docs.travis-ci.com/user/reference/windows/

For the moment there is one package missing on our side, and that's the
Windows Subsystem for Linux. This is only available in very up-to-date
builds of Server 2016, so the Windows update process must be run before
this can be applied. I need to come back and enable the updates for both
the Windows boxes.