feat/role: use Git Bash as Windows shell
The way to do this is to set the Git Bash location at the beginning of
the PATH variable. Jenkins has a 'tool location' section where you can
specify this, but you can only use this if *all* the agents are Windows,
because it's a system wide setting.

See the answer here:
https://stackoverflow.com/questions/35043665/change-windows-shell-in-jenkins-from-cygwin-to-git-bash-msys