feat/role: install Rust on Windows
This uses a role rather than just the script because the install really
needs to run as the user who is executing the Jenkins slave service
process. The installer will setup user-based environment variables that
make Cargo available on the command line.

The Windows machine also needs to be restarted after these and other
environment variables are added. Changes to the system PATH that are
made by the Jenkins slave role don't seem to take effect until after
restarting.