feat/vm: provision osx slave
A playbook that sets up an OSX machine with an installation of Java and
a Jenkins user with a public key. I've verified that this setup allows
Jenkins to establish an SSH connection and add the machine as a node.
For now the inventory contains the the internal IP address of the slave
rather than its external IP, as we're having some connectivity issues
with that.