refactor/role: rust compatibility for OSX
Provides some changes for getting the Rust role to run on OSX. The main
issue here was running into a problem with temporary files that Ansible
uses. This is why the ANSIBLE_PIPELINING variable is set. See here for
the issue: https://docs.ansible.com/ansible/latest/user_guide/become.html#becoming-an-unprivileged-user

This could potentially be problematic in the future, because the copy
and template modules don't work with pipelining, and that's a huge
limitation. For now though, these aren't necessary.