feat/env: provision Jenkins on AWS
This adds a development setup for running Jenkins on AWS. I've used 2
Linux hosts just so I could prove it would work if there was more than 1
slave in the mix (which is likely to be the case for a non-local setup).

The dynamic inventory seemingly needs to consist of the EC2 script plus
a static inventory file that groups the machines based on their
hostnames. The EC2 dynamic inventory maps the hostnames in the static
inventory to the Name tag on the EC2 machine.

The Jenkins CASC file is updated to use a loop to dynamically populate
all the Linux machines in the 'slaves' group. This will probably need to
be split into a 'linux_slaves' or 'docker_slaves' and 'windows_slaves'
groups for when Windows support is added shortly. It will also need to
have an if condition to distinguish whether we're running a local setup
or not, since it references properties that will only be available on
EC2, e.g. 'ec2_public_dns_name'.