fix/role: correct git installation
Even although this provision is seemingly running as the Administrator
user, it still wouldn't let Git install, complaining about permissions.

There is a thread and some resolutions to the issue here:
https://github.com/chocolatey/choco/issues/1048

I chose to just assign FullControl permissions to the Chocolatey
installation directory. This issue doesn't occur on the local VM and I
suspect this change will be compatible with that.