refactor/vm: split instance up and Ansible run
These had to be separate out because Jenkins has a slave configuration
and all the slaves machines need to be available at the point when the
master is provisioned. This is because the list of slaves is provided by
the Ansible dynamic inventory, and the Windows slaves wouldn't be
available if we spun them up after the master provision.

On the other hand, the Windows slave provision requires a running
Jenkins instance to connect to, so we spin the Windows machine up but
don't do the Ansible run until after the Jenkins master has been
provisioned.

The 2 scripts need to share information, so the instance ID and
generated passwords are written out to a hidden directory. It's not a
security issue to store this password on a local machine, as this is
just a development setup.