feat/vm: slave config based on environment
The jinja2 template for the Jenkins configuration uses a variable to
determine if we're running in cloud mode or not. In that case the slaves
are populated using a different method that queries the inventory. It's
necessary for there to be different paths, because some of the variables
that are injected by dynamic inventory don't exist in the local setup.