feat/vm: wireguard roles
There is a role that installs wireguard (common to both client and
server) and one for both client and server. The client and server roles
only really contain configuration.

There is a wireguard group that contains the variables that are shared
between both the client and server configurations. This can be applied
to any playbooks we want to apply the roles too.