feat/env: wireguard sandbox environment
Sets up a little sandbox WireGuard environment between 2 machines,
one being a local VM and the other being on AWS.

With this setup the 2 machines are able to ping each other, and you can
also SSH from one machine to the other. I did see some issues with the
AWS machine not being able to ping the local VM after an SSH attempt
failed, but it seems to start working again when the local VM pings the
AWS machine. Will need to keep an eye on it.