feat/env: switch to using ubuntu for wireguard
There are issues using WireGuard with CentOS 7.5 because that has a
fairly old kernel version (3.10), and if you're going to use Docker
with WireGuard, you're better being on a more up-to-date kernel, since
Docker uses the same kernel as the host machine. The WireGuard utilities
for Docker depend on installing OS header files on the host, so the
container and host kernel versions generally need to match.

Ubuntu has a pretty recent version of the kernel, so it seemed easy to
change to that. We'll need to upgrade all the other machines in this
estate, but first I want to get the WireGuard sandbox working.