feat/env: use wireguard on jenkins master
The Jenkins master machine gets changed to Ubuntu 18.04 to make it
easier to use with WireGuard, since it has a much more up-to-date kernel
version than CentOS 7.5.

The wireguard and wireguard-server roles are then applied to the
jenkins-master playbook. However, WireGuard isn't actually running on
the host machine, but rather, in a container that's linked the Jenkins
master container. The docker-compose definition has been changed to add
the WireGuard container and establish this link. The postfix container
was temporarily removed from this setup until I figure out how to link
all 3.