feat/vm: remove route on macos wireguard setup
Without the removal of this route, the machine will be able to connect
to the VPN, but it won't be able to connect to the internet.

There is also a bit of basic service management here. Need to configure
it to run as a service on macOS, but that will be provided later.