refactor/role: improve use of variables
Uses better variable names to describe all the IP addresses in use.

It also changes the IPs to match those used by the streisand repository,
which was done when trying to debug problems with macOS trying to
connect, but the addresses are arbitrary anyway, so I can just leave
them as is.