feat/env: initial production infrastructure
The first attempt to describe the production infrastructure. We have a
VPC with a public subnet and a private subnet. The Jenkins master will
be in the public subnet and the slaves will be in the private subnet.
Eventually we'll be able to remove the definition of the slave machine
from this setup because we'll use the EC2Plugin for Jenkins that will
spin up the slaves dynamically.

The rules for the security groups setup the communication between the
public and private subnets. The rules need to be declared using
`aws_security_group_rule` resources, as opposed to using the
`aws_security_group` resource, because the groups reference each other,
and this creates a cycle in Terraform's acyclic graph.