feat/vm: machine for running ansible
This introduces a machine where we can run Ansible provisions from,
given that the other machines don't have direct SSH access. It seems
that communication between machines in different security groups on the
same subnet also need the rules to establish connectivity between the 2,
as with the public/private subnet communication. I don't understand
this, but they are definitely necessary for things to work correctly.