feat/vm: ansible provisioner for prod infra
Adds some roles for setting up a machine to act as an Ansible
provisioner inside the VPC. This obviously needs to run from a machine
that already has Ansible on it. That could be a developer laptop or some
internal machine we eventually connect to Jenkins.

Right now this adds a private key to the Ansible Bastion host. This
should be removed and we should setup some kind of SSH tunnelling to
avoid having private keys on the Bastion host. For now we need this
setup for developing the prod infrastructure.