fix/script: update based on distribution
On Ubuntu machines we also install Python, since newer versions don't
have an install by default.

There is also a sleep in the Makefile before provisioning the Bastion
host, as a general yum update runs before creating the ansible user.
This can be changed later to use a loop that tries to establish an SSH
connection.