fix/env: use private dns name for prod slaves
The prod slaves will not have public DNS names, so we use the internal
field instead. This also temporarily disables the Windows host as that's
not available yet.

Really we need two different configurations for the prod and dev
environments, or some other kind of way for the details to vary, as the
private hostnames may not work in the dev setup. I've added a task to
look at this later.