fix/vm: install python libraries in virtualenv
The pypsrp library wouldn't work correctly with a system-wide
installation because it requires a specific version of the requests
library. Ansible also needs to be installed in the virtualenv because
the ec2.py script references Ansible Python modules.

It's up to the user doing this provisioning to activate the virtualenv
before running the provision. We might automate this at some point in
the future, but don't want to invest time in this now.