feat/env: debugging for the prod environment
This introduces a debugging configuration for the production
environment. If you set DEBUG_JENKINS_ENV to true before calling the
target to create the environment, a Windows Bastion host will be
included in the setup. This will allow you to RDP to the Windows machine
on the private subnet. It is controlled with a count variable that is
set to 0 in the default configuration and 1 in the debugging
configuration.