docs/vm: describe the generally useful machines
There are a lot more VMs available in the Vagrantfile, but most of the
others have very specific purposes, for example a WireGuard sandbox, so
I don't think they're worth listing here.

I've also removed references to specific applications in the table,
because really these machines are quite generic and could be used for
building most Rust applications, rather than just SCL and Crust
specifically.

For much the same reason, I've also removed the section on
'Building Crust', because I don't think this repository should be
targetted to that specifically, and the instructions for building it
should probably just be contained in that repository, otherwise we'd
need to maintain 2 documents.