feat/env: slave management with ec2 plugin
This configures the Jenkins EC2 plugin for dynamically spinning up
slaves. There is a new `jenkins-ec2` user created just for this purpose,
and it's locked down to only the operations it needs for this.

This removes the need for the static slaves. So far it's working for
Linux on the default-VPC. The next step is to change to the production
VPC and launch on the private subnet.