refactor/jenkins: use static windows slave
We have agreed that it will take some more time to figure out how to get
the EC2 plugin to work with Windows. You can't really use a static slave
agent service because it has a reference to a Jenkins master, which
would need to be injected in every time the machine is spun up. I think
the EC2 plugin does its own agent management.

This also introduces a wrapper script for provisioning the Windows
production instance, to get round the problem with the dollar in the
password.