fix/vm: remove mingw package
This has decided to randomly stop working today, complaining it can't
retrieve the package from the URL. Seems to work fine when you run it
manually. I tried to replace it with the msys2 package but that seemed
to be taking over 15 minutes to install, so I'm not sure if it's started
prompting for user input or something.

We can add this back in later. It will stop SCL from building
successfully, but it won't stop the slave coming up.