refactor/env: describe dev infra with terraform
Switches from using Vagrant machines and ad-hoc scripts to Terraform for
describing the infrastructure for the dev environment. This is much
better because Terraform is idempotent and the Vagrantfile is getting
quite crowded as it is.

For the dev environment there are 4 machines:

* Jenkins Master
* Docker Slave 1
* Docker Slave 2
* Windows Slave

The Mac is still out of action until we get the VPN working.