refactor/vm: get password before provision
Changes the Windows machine provisioning for the dev environment to
retrieve the password at the same time it's going to run Ansible. By
this time the password should be available on the instance, so the loop
won't really have much effect. It was previously used when the
password was being retrieved after the instance was spun up.

Since we aren't using Vagrant for this setup any more, we need to
retrieve the hostname of the Jenkins master via the awscli.