refactor/env: change master and slave configs
Changes the docker-compose definition and the configured slaves based on
the type of environment we're running.

In the local and the AWS dev setups, we don't need WireGuard installed,
so this needs to be removed from the compose definition. It actually
doesn't work in these setups (although I'm not sure why), so it had to
be removed, otherwise we could have just left it in and it wouldn't have
done too much harm.

The dev environment will just use a couple of static Linux slaves, so we
remove the use of the EC2 plugin and loop over the Linux slaves group,
as it was before.