refactor/env: improve prod env process
I've been working with this setup enough such that the manual steps were
slowing me down, due both to errors in the steps and because they take a
bit of time.

The beginning of the process will prompt you for your AWS keys and the
vault password, and these will then be setup on the Bastion host. It
will also copy the `jenkins_env` key to the Bastion, as this is required
to get the password for the Windows instance.

It should be noted that when we provision the infrastructure for real,
we should remove all of this private information from the Bastion host.
When we can get the proper infrastructure setup, we can do an SSH tunnel
to avoid having to put any keys on the Bastion.