feat/box: packer templates for 2 machines
These are Packer templates for the Docker slave and the Jenkins master.
The boxes build just fine, but unfortunately when you update the VM to
use the pre-baked Jenkins master, you can't connect to port 8080 or
50000 on the VM. I tried for a good couple of hours to debug the issue
but couldn't resolve it. I suspect it may have something to do with
Docker, but I'm not entirely sure.

I think it's worth leaving the templates in here anyway, as we might be
able to get it working at some point in the future.