refactor/env: provision dev and prod together
The dynamic inventory settings we were using didn't permit the
provisioning of 2 different environments at the same time, due to
machines having the same Name tag. You have to add another filter for
dynamic inventory that allows it to distinguish which environment the
machine is in.

This adds the ini file for each environment with the appropriate
environment tag set. Also refactored one of the scripts to pass the
environment as an argument rather than having 2 different scripts for
dev and prod.