refactor/env: use 2 different ini files
We need to have 2 different ini files for the production environment,
because the first initial provision takes place outside EC2 on the host,
and the 2nd provision takes place inside EC2 on a Bastion host. The
first needs to use the public IP addresses of the machine and the 2nd
needs to use private IP addresses. It's simpler and really not a lot of
overhead to just have 2 files. The relevant scripts have been refactored
to allow passing the name of the file.