feat/jenkins: use github authentication
For the prod environment we're now going to be using GitHub as the
authentication mechanism. We have an GitHub app setup that enables the
integration.

The security matrix has been setup so that the QA team have admin rights
and developers have the rights they need for running and viewing jobs,
but not being able to edit them (since they're defined in code and we
want to keep them that way) or adminster the Jenkins setup (which is
also defined in code).