fix/jenkins: simplify security matrix
When setting GitHub as the security realm I was getting stuck in an
authorization loop when I tried to log in. GitHub kept on asking me to
authorize the application we're using to integrate GitHub and Jenkins.
The log was complaining about a bad cookie, but I suspect this isn't the
actual reason. If there was something wrong with the security matrix,
perhaps somehow that could have caused the login loop.

This uses an ultra simple security matrix that just sets myself as the
admin.