feat/vm: extra disks for docker slave
The template for the Docker slave is extended to provide extra disks.
One is for the Docker storage and the other for the Jenkins workspace.
As per this commit, the Docker storage isn't yet configured.

We also add an AWS machine to the Vagrantfile that allows us to
experiment and configure Docker for production storage on the mounted
volume.