feat/vm: jenkins slave to use extra disk
Configures the Jenkins slave to use the external disk for its working
directory. The disk needed to be formatted after the jenkins user
already exists (so that the jenkins user can be set as the owner of the
directory) so the order in the roles in the playbook have changed a
little bit.