feat/vm: install libsodium on windows
It turned out that with version 1.0.16 of libsodium, the mingw archive
for it had been removed from the download location. I installed this
version, but then found out there were plans to upgrade to 17, so I'm
working with that version instead.

I think it's easier to just pre-install this library to a known
location and hit the unreliable web server during the provisioning
process rather than when building.