refactor/env: remove sleep in favour of wait
A little script is introduced to wait on SSH becoming available, after
which it runs an apt or yum update. Previously this was in the user data
script and so we had no visibility on when or if it completed. We were
just using a sleep and hoping things worked.

Now we can avoid the sleep, which is particularly beneficial for running
the provision again, as there's no waiting around when things have
already been setup.