fix/vm: use private ip address for slave port
I suspect the reason this isn't connecting is because
jenkins.maidsafe.net is now pointing at the proxy, which doesn't have
port 50000. I guess another alternative to this would be to forward the
traffic on the proxy, but not sure how it deals with non HTTP requests.

I'm not certain this will fix the issue, but it's got a good shot.