refactor/script: use image name to determine OS
This script wasn't really generic enough for use with the dev and prod
environment at the same time.

It's now been changed to detect the OS, using the AMI ID from the
machine description. Waiting for SSH to come online also didn't quite
work correctly. Previously this script was being used for the prod
environment and it turned out just by coincidence that the machines were
always ready, because the prod environment takes much longer to
provision.