feat/role: simple letsencrypt configuration
For now this is just installing certbot and adding some configuration
for redirecting to the LE webserver. It seems that this may need to be
an *initial* configuration, because according to the guide I'm reading,
when you're generating the certificate, you can't have HAProxy listening
on both 80 and 443 at the same time.