feat/role: haproxy configuration for ssl
This is based on the configuration detailed here:
https://serversforhackers.com/c/letsencrypt-with-haproxy

This article details that you can't have HAProxy listening on port 443
when the certificate is being generated initially, and that's why this
role exists separately.

The process for bringing up the environment is:

* Provision the HAProxy machine and start the service only running on 80
* Provision the Jenkins Master machine and start the service
* Generate the certificate
* Copy the new HAProxy configuration with port 443 and SSL configured

Due to having 2 different configurations for the proxy, and the fact
that the certificate generation needs to happen *after* Jenkins is
started, we have this setup in a separate role.