feat/box: ubuntu docker slave box
The idea was to change the Docker slave from CentOS to Ubuntu. For the
AWS AMIs this has been fine, but as it stands currently, this box isn't
working. However, I want to keep the changes anyway, as I think it's
quite close to working.

The problem was getting Ansible to run against the box. The original
idea was for the Ansible run to take place from the host (and this
worked with the CentOS machine), but Packer was generating a bizarre
inventory file that had a strange port assigned for SSH, and hence
Ansible wasn't able to connect. I'll come back to this later.