fix/vm: docker slave disk issue
I'm not actually sure why I changed this disk mount to be owned by the
root user when it actually needs to be the Jenkins user who writes to
that location. The commit where it was changed unfortunately doesn't
have an explanation. This should fix the issue, and a new AMI has been
generated.