feat/box: template for windows ami
Provides a template for building a Windows Server 2016 AMI. We wanted to
add a 2nd Windows slave to our setup, but the problem with that was that
you get a randomly generated password for the Administrator user, and
this would prevent doing an automated Ansible run across a set of
Windows machines, because each of them have a different password, and
Ansible assumes a uniform authentication method for connecting to all
machines in a group.

You are advised to run the EC2 launch scripts as part of an AMI
generation, and these are the scripts that set the password for the
Administrator user. Rather than interfere with this, I originally
decided to opt for adding an ansible user that I could provide with a
known password that could then be used in the Ansible run. However, this
didn't work, because for a reason I couldn't determine, after spinning
the AMI up, I couldn't authenticate as the ansible user. It seems at
some point the password for this user was also being reset.

For this reason, the template here is quite simple. It just formats the
additional disk and installs a bunch of packages with Chocolately. These
are packages that normally take quite a bit of time during an Ansible
run, so using this AMI will definitely save us quite a bit of
provisioning time.

The restart needs to occur before the EC2 launch scripts because some of
these Chocolately packages require the machine to be restarted, and the
'sysprep' script was displaying a GUI to this effect.