feat/role: create ansible user on windows
This role is being used to create the user because creating the user in
an AMI didn't seem to work correctly. The password wasn't set for the
user after the AMI is spun up. This playbook is ran as part of an
environment provision to create and set an Ansible user on each slave
that has the same password. That means we can run Ansible uniformly
against all the slaves.