fix/vm: use correct slave reference
For the cloud environments, each Windows slave needs to reference its
own secret key for connecting to Jenkins. It uses the full name assigned
to the full name tag on EC2, which Ansible makes available via dynamic
inventory.

Also changes the working directories for the slaves to use the
additional D drive.