feat/env: optional cache clearance
I have found the dynamic inventory cache to be extremely unreliable, so
I generally just clear it. However, it is extremely slow when it's
cleared, so I want to just disable clearing it for this one case where
the ansible user password gets set on the Windows machine. This has to
be done for all Windows machines sequentially, so it's very slow to
clear the cache and wait for each one.