refactor/env: upgrade jenkins and proxy instance
A micro instance is just a bit too small for the Jenkins master. It's
been observed to have crashed a few times after running for a day or 2.
A bit of extra memory should hopefully take care of this.

Changing the proxy from a t2.micro to a t3.micro gives it another CPU
core, which should be good for servicing requests, even if they are
quite small.

Also terminate volumes for the Jenkins master.