fix/env: jenkins master as t2.medium instance
We tried to use a t3.small instance for the Jenkins Master, but it turns
out that disk management is completely different for t3 types.

They are assigned different block device names, but unfortunately the
device names are also not assigned in a reliable order. There is an
issue about it on the Terraform repository:
https://github.com/terraform-providers/terraform-provider-aws/issues/4864

There are links to some solutions on that post, but I don't have time to
investigate just now. We have agreed for now to just use a t2.medium
instance.