refactor/env: prod env parity with staging
This changes the prod environment as per the staging environment, to
remove references to the environment name in variables and definitions
of types. Now whenever changes are made to prod, we should be able to
just copy the file over to staging and add in any new variables that
were created. This seems to be the best option for maintenance at the
moment.

It turned out also that the prod environment was using the wrong public
key for the jenkins keypair. That's been fixed here.