feat/env: remove 2nd docker and windows slaves
I originally added these to the dev environment because I wanted to
prove that you could populate the list of slaves in the CASC
configuration using a loop over all the slaves in the inventory. That
concept is now well established, so I'm removing the 2nd slaves to keep
the cost down for the environment.