feat/env: initial setup for qa environment
This provides the initial infrastructure definition, and the Ansible
environment is copied from staging, so things like the WireGuard IPs
will need to be changed.

The QA environment is different from staging/prod in one detail. There's
no URL for QA because we hit a limit on with elastic IP allocations. We
can possibly put in a support request to get this increased, but for now
we'll just go with the EC2 DNS name.