feat/env: provisioning for qa environment
The provisioning for QA is almost identical to staging and production,
except for the fact that it doesn't have a URL, so we'll just use the
EC2 DNS name for the proxy machine. There are some modifications to the
scripts and there's also a new script introduced for doing the SSL
configuration (previously it was just done directly in the Makefile).