fix/vm: service starting issue
On a re-provision the Jenkins slave service was failing to start. The
reason for this was because the service user had the service logon
privilege removed. It seems to be the case that the win_user_right
module actually reverses the permissions when it runs again, so I've put
a condition on this.