fix/script: assign qa url
This was another special change I made for the QA environment before we
had our limit increased to get more elastic IPs. It can be removed. It
didn't effect anything because it was only be referenced by the wg
setup, which can reference either the EC2 DNS name or our custom DNS.