fix/play: remove or conditional
The "none" environment was for compatibility with the local setup, but
it would actually have to be a logical and. Rather than changing it, I'm
just removing these, because the local environment hasn't been looked at
for a long time. It would need some work to get the local environment
running again, so for now it won't really be supported as a scenario.