feat/env: initial attempt at util slave
This is a slave that's going to be used for deployments and such things.
The difficulty with it is going to be the security group configuration.
This is the first initial attempt, which is opening a connection between
the Bastion and the util slave (which will be removed) and trying to
establish internet connectivity.