refactor/env: move docker_slave_ssh_public_key
This moves this variable from the `all` group to the `linux_slaves` and
`osx_slaves` groups. This variable also really needs to have a more
generic name, like `jenkins_slave_ssh_public_key`, because it won't be
specific to the Docker slave any more.

The reason why it's being moved is because the `util_slaves` group needs
to provide a different value for this key.