feat/box: packer template for util slave
Provides a template and a target for running the template.

In the provisioners, you can see some strange things going on
with the installation of Python. For some reason Ubuntu 18.04 doesn't
ship with an installation of Python, but for some unknown reason,
sometimes when you try to install it here, it complains that there's no
package available for it. This is the reason for having the sed command
that edits the repository sources. Sadly though, this doesn't appear to
fix it all the time. You need to just keep running the target until it
works.

If the problem continues to persist, we can create a base AMI that just
has Python installed.