refactor/role: docker-slave-images
Changes the role from just pulling in docker images to also logging in
to Dockerhub so that the machine would be able to push an image.

There may also need to be some other things added at some point in the
future, and it feels better to put this functionality in here rather
than the jenkins-slave role, as you wouldn't necessarily want to login
or even have Docker installed on a slave.