feat/box: new AMI ID for docker slave
I'm still experiencing issues with the python and python-pip packages
being available on Ubuntu Bionic. I've now modified the scripts to use a
loop for installing them. I don't know if this actually works because
after I added these, the first time I ran it there were no issues with
the packages! So we will see the next time the box is generated.