feat/env: new terraform definitions
This isn't really a new environment, but more a place to put things that
are common to all environments, for example, an S3 bucket for build
artifacts, which is what this code does. We don't really need one
of these buckets per environment; the same bucket and user is fine for
all.

This also uses the S3 backend for Terraform state.