fix/env: jenkins inaccessible during reprovision
There was an issue with Jenkins going offline during the reprovisioning
process. This was due to HAProxy not forwarding the request, due to a
bad configuration.

There are 2 different configurations for HAProxy: one in this role and
one in the haproxy-ssl-config role. If you look at the first commit for
the haproxy-ssl-config role, it discusses why, so I won't repeat that
here. During the reprovisioning process, the reason why HAProxy was
going offline is because the first config file is overwriting the
second one, which it doesn't like because the first config file doesn't
have SSL enabled.

The issue has been resolved by providing a 'run_mode' variable that
indicates whether we're doing an initial provision or a reprovision. The
first config file will only be copied during the initial run.