refactor/role: split ansible-provisioner
Splits the ansible-provisioner out into 2 roles: one that installs
Ansible and related libraries, and another that sets up all the keys
necessary for the ansible user on the Bastion host. The role that sets
up the context for the user on the Bastion host isn't really thinking
about automation, since the machine was being set up to be used
manually. That's why I've specifically put 'bastion' in the role.

These should have probably been decoupled right from the start.

The role for installing Ansible is now also being used by the util
slave.