feat/env: new IAM user for building AMIs
As the comment in the code says, this was taken from the Packer
documentation. So far I've tested that the box generation can run with
this user, but there's been a problem with the Ansible run, so I can't
confirm whether these permissions work end to end, but I suspect they'll
be fine.