fix/role: change docker login mechanism
Unfortunately when I was trying to generate an AMI from the util slave,
there was a weird error coming from the docker_login module. The same
error is detailed here:
https://github.com/ansible/ansible-modules-core/issues/5685

Trying it just using the 'manual' method. It works for using Vagrant to
bring up the util slave, but I can't tell yet whether it'll work for
generating the AMI from the slave.