fix/box: docker login with shell command
Unfortunately using the direct shell script for the docker login still
didn't work when running Packer from the util slave. It's complaining
about not being able to access the Docker daemon, despite the fact that
the jenkins user is in the docker group.

I'm going to try the login from a shell script rather than Ansible.