fix/box: docker login as shell command
As the comment on the commit message says, for some unknown reason, when
this role is executing from the util slave in the context of a Packer
run, it complains about not having access to the Docker daemon (despite
the fact that the user it's running as has been added to the docker
group). We can copy the password over in a file and then just run a
docker login command directly with an inline shell script, and this
seems to work.