refactor/box: use env variables for settings
With the util slave on the private subnet, it won't work with the
default Packer configuration, which is to launch the machine into the
default VPC and create an ad-hoc security group. The template now
specifies that the VPC, subnet and security group IDs be supplied as
environment variables. This can be done using global environment
variables in Jenkins.

The only problem with this setup is that it won't work for local builds
from a laptop. Can possibly address that by using a different target in
the Makefile.